requirements.txt
astroid==1.6.5
dj-database-url==0.5.0
Django==2.0.6
djangorestframework==3.8.2
gunicorn==19.8.1
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
psycopg2-binary==2.7.5
pylint==1.9.2
python-decouple==3.1
pytz==2018.5
six==1.11.0
whitenoise==3.3.1
wrapt==1.10.11
